requirements.txt
Django==3.1.3
django-filter==2.3.0
djangorestframework==3.11.0
django-cors-headers==3.5.0
drf-yasg==1.17.1
coverage==5.1
gunicorn==20.0.4
psycopg2==2.8.6
pytest==6.1.2
model-mommy==2.0.0
flake8==3.8.4
drf-yasg==1.17.1
